built a basic learn frontend page
what has been done?
1. Building a basic structure for the learn frontend page by adding the navbar, introduction and a few other sections

What needs to be worked on?

1. I still need to create an HTML , CSS resources page and debugging a simple blog page
2. I'm planning on adding more pages like different CSS frameworks like bootstrap, tailwind CSS etc

What this page needs help on?

I'm not very proficient in Javascript enough to make another subpage about it and its frameworks so it would be helpful to find someone who could work on its structure. I dont mind coding the webpage so long as someone who is experienced in JS provides a markdown file on it